backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
asyncio